fixed errors with notes crud

+++ src/main/java/dev/bego/laika/notes/NoteController.java
@@ -1,9 +1,12 @@
 package dev.bego.laika.notes;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,21 +16,29 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import dev.bego.laika.users.User;
+
 @Controller
 @RequestMapping(path = "${api-endpoint}/notes")
 public class NoteController {
-    
-    private final NoteService noteService;
 
     @Autowired
-    public NoteController(NoteService noteService) {
-        this.noteService = noteService;
+    private NoteService noteService;
+
+    @GetMapping("")
+    public ResponseEntity<List<NoteDto>> getAllNotesByUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        List<NoteDto> notes = noteService.getAllNotesByUserId(currentUser.getId());
+        return ResponseEntity.ok(notes);
     }
 
-    @GetMapping("/user/{userId}")
-    public ResponseEntity<List<NoteDto>> getAllNotesByUserId(@PathVariable Long userId) {
-        List<NoteDto> notes = noteService.getAllNotesByUserId(userId);
-        return ResponseEntity.ok(notes);
+    @PostMapping("")
+    public ResponseEntity<NoteDto> createNote(@RequestBody CreateNoteDto noteDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        NoteDto createdNote = noteService.createNoteForUser(currentUser.getId(), noteDto);
+        return ResponseEntity.ok(createdNote);
     }
 
     @GetMapping("/{noteId}")
@@ -37,20 +48,32 @@ public class NoteController {
                 .orElse(ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/user/{userId}")
-    public ResponseEntity<NoteDto> createNote(@PathVariable Long userId, @RequestBody NoteDto noteDto) {
-        NoteDto createdNote = noteService.createNoteForUser(userId, noteDto);
-        return ResponseEntity.ok(createdNote);
-    }
-
     @PutMapping("/{noteId}")
     public ResponseEntity<NoteDto> updateNote(@PathVariable Long noteId, @RequestBody NoteDto noteDetails) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        Optional<NoteDto> note = noteService.getNoteById(noteId);
+        if (!note.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        if (note.get().getUser().getId() != currentUser.getId()) {
+            return ResponseEntity.badRequest().build();
+        }
         NoteDto updatedNote = noteService.updateNote(noteId, noteDetails);
         return ResponseEntity.ok(updatedNote);
     }
 
     @DeleteMapping("/{noteId}")
     public ResponseEntity<Void> deleteNoteById(@PathVariable Long noteId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+        Optional<NoteDto> note = noteService.getNoteById(noteId);
+        if (!note.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        if (note.get().getUser().getId() != currentUser.getId()) {
+            return ResponseEntity.badRequest().build();
+        }
         noteService.deleteNoteById(noteId);
         return ResponseEntity.noContent().build();
     }
